reformat spacing, DRY code, check passthru value only once, use call().
The spacing seemed to be inconsistent when looking on Github as well
as in Atom. It seemed spacing was at 4 so I used that.

Instead of writing very similar code for both functions I used a
single function and passed in the stream to hook. (DRY)

Instead of checking the value of "addl" (I renamed to "passthru")
every time a write is called, I check it once and set the write()
to do what's desired every time. This allows simply setting the
provided callback as the write function when "passthru" is false.

switch to call() from apply() to avoid using arguments.